+++ RedEye/Engine/RE_ShaderImporter.h
@@ -14,7 +14,7 @@ namespace RE_ShaderImporter
 	bool LoadFromBinary(const char* buffer, size_t size, unsigned int* ID);
 	bool GetBinaryProgram(unsigned int ID, char** buffer, int* size);
 
-	bool Compile(const char* buffer, unsigned int size, unsigned int GLCompile = 0x8B31);
+	bool Compile(const char* buffer, size_t size, unsigned int GLCompile = 0x8B31);
 
 	//get last error
 	const char* GetShaderError();
